fix: normalize phone value before cleaning
The previous logic caused a weird behavior when applied to Arabic letters or some combinations of letters (eg: ي), causing the output of the replace function to convert the string to a faulty bitstring that does not correspond to any string. We can simply reproduct that by invoking `String.replace("ي", ~r/[^0-9\s()\-٠١٢٣٤٥٦٧٨٩+]|(?<status)\+/, "")` or even `String.replace("ي", ~r/[٠]/, "")`.